Flatten Envs are integer-indexed arrays of local variables.
At compile-time, we register all the `LocalName`s used in a method
body, and assign integer indices to them. At run-time, we allocate
an array of the right amount of local variables at the beginning of
the method, then mutate it from the body.

Captures are stored in a separate, immutable array, so that they do
not need to be copied every time we enter a method with captures.